Isolate buffer_selector map
